requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==5.0.1
attrs==24.2.0
bcrypt==4.2.1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cloud-sql-python-connector==1.14.0
cryptography==44.0.0
dnspython==2.7.0
email-validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.5
fastapi-cli==0.0.5
frozenlist==1.5.0
google-api-core==2.23.0
google-auth==2.36.0
google-cloud-core==2.4.1
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
idna==3.10
jinja2==3.1.4
markdown-it-py==3.0.0
markupsafe==3.0.2
mdurl==0.1.2
multidict==6.1.0
passlib==1.7.4
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.0
pyasn1==0.6.1
pyasn1-modules==0.4.1
pycparser==2.22
pydantic==2.10.0
pydantic-core==2.27.0
pydantic-settings==2.6.1
pygments==2.18.0
pyjwt==2.10.1
pymysql==1.1.1
python-dotenv==1.0.1
python-multipart==0.0.17
pyyaml==6.0.2
requests==2.32.3
rich==13.9.4
rsa==4.9
shellingham==1.5.4
sniffio==1.3.1
sqlalchemy==2.0.36
sqlmodel==0.0.22
starlette==0.41.3
typer==0.13.1
typing-extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
uvloop==0.21.0
watchfiles==0.24.0
websockets==14.1
yarl==1.18.3
